Location stuff works

+++ app/src/main/java/com/androidimpact/app/activities/EditIngredientLocationsActivity.java
@@ -2,29 +2,44 @@ package com.androidimpact.app.activities;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.recyclerview.widget.ItemTouchHelper;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
+import android.app.Activity;
+import android.content.Intent;
+import android.os.Bundle;
+import android.util.Log;
+import android.view.View;
+import android.widget.Button;
+import android.widget.EditText;
+
 import com.androidimpact.app.Location;
 import com.androidimpact.app.LocationAdapter;
 import com.androidimpact.app.R;
+import com.google.android.gms.tasks.OnFailureListener;
+import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.material.snackbar.Snackbar;
+import com.google.firebase.firestore.AggregateQuerySnapshot;
+import com.google.firebase.firestore.AggregateSource;
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
 
-import android.os.Bundle;
-import android.util.Log;
-
 import java.util.ArrayList;
 
-public class EditIngredientLocations extends AppCompatActivity {
+public class EditIngredientLocationsActivity extends AppCompatActivity {
+
     final String TAG = "EditIngredientLocations";
 
     RecyclerView locationRecyclerView;
     ArrayList<Location> locationArrayList;
     LocationAdapter locationViewAdapter;
+
+    // Views
+    Button addLocationBtn;
+    EditText newLocationInput;
 
     // Firestore
     FirebaseFirestore db;
@@ -35,9 +50,14 @@ public class EditIngredientLocations extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_edit_ingredient_locations);
 
+        getSupportActionBar().setTitle("Edit Locations");
+
         // initialize Firestore
         db = FirebaseFirestore.getInstance();
         locationCollection = db.collection("locations");
+
+        // initialize views
+        newLocationInput = findViewById(R.id.location_editText);
 
         // initialize adapters and custom
         locationRecyclerView = findViewById(R.id.location_listview);
@@ -81,17 +101,36 @@ public class EditIngredientLocations extends AppCompatActivity {
 
                 Log.d(TAG, "Swiped " + location + " at position " + position);
 
-                // delete item from firebase
-                locationCollection.document(id)
-                        .delete()
-                        .addOnSuccessListener(aVoid -> {
-                            Log.d(TAG, location + " has been deleted successfully!");
-                            Snackbar.make(locationRecyclerView, "Deleted " + location, Snackbar.LENGTH_LONG).show();
+                // we first count up how many collections there are
+                // if there is only 1 left, we cannot delete
+                locationCollection.count().get(AggregateSource.SERVER)
+                        .addOnSuccessListener(aggregateQuerySnapshot -> {
+                            long count = aggregateQuerySnapshot.getCount();
+
+                            if (count <= 1) {
+                                makeSnackbar("You need at least 1 location!");
+                                // "swipes" the element back lmao
+                                locationViewAdapter.notifyItemChanged(position);
+                                return;
+                            }
+                            // delete item from firebase
+                            locationCollection.document(id)
+                                    .delete()
+                                    .addOnSuccessListener(aVoid -> {
+                                        Log.d(TAG, location + " has been deleted successfully!");
+                                        makeSnackbar("Deleted " + location);
+                                    })
+                                    .addOnFailureListener(e -> {
+                                        makeSnackbar("Could not delete " + location + "!");
+                                        Log.d(TAG, location + " could not be deleted!" + e);
+                                    });
                         })
                         .addOnFailureListener(e -> {
-                            Snackbar.make(locationRecyclerView, "Could not delete " + location + "!", Snackbar.LENGTH_LONG).show();
-                            Log.d(TAG, location + " could not be deleted!" + e);
+                            Log.d(TAG, "Failed to count locations!", e);
+                            makeSnackbar("Failed to count locations! Check the logs... ");
                         });
+
+
             }
             // finally, we add this to our recycler view.
         }).attachToRecyclerView(locationRecyclerView);
@@ -111,10 +150,7 @@ public class EditIngredientLocations extends AppCompatActivity {
             for(QueryDocumentSnapshot doc : queryDocumentSnapshots) {
                 String id = doc.getId();
                 try {
-
-                    String location = doc.get("location", String.class);
-                    Location l = new Location(location);
-                    locationArrayList.add(l);
+                    locationArrayList.add(doc.toObject(Location.class));
                 } catch (Exception e) {
                     Log.i(TAG + ":snapshotListener", "Error retrieving document " + id + ":" + e);
                     errorCount += 1;
@@ -127,5 +163,54 @@ public class EditIngredientLocations extends AppCompatActivity {
             Log.i(TAG, "Snapshot listener: Added " + locationArrayList.size() + " elements");
             locationViewAdapter.notifyDataSetChanged();
         });
+    }
+
+    /**
+     * This is run whenever `R.id.addLocationBtn` is pressed
+     */
+    public void locationBtnPressed(View view) {
+        Log.i(TAG + ":locationBtnPressed", "Adding a new location!");
+        String locationName = newLocationInput.getText().toString();
+
+        // return fast on empty location
+        if (locationName.equals("")) {
+            makeSnackbar("Please enter a location!");
+            return;
+        }
+
+        Location l = new Location(locationName);
+        Log.i(TAG, "Adding location " + l.getLocation());
+        newLocationInput.setText("");
+        locationCollection.document(l.getId()).set(l)
+                .addOnSuccessListener(unused -> {
+                    makeSnackbar("Added " + locationName);
+                })
+                .addOnFailureListener(e -> {
+                    makeSnackbar("Failed to add " + locationName);
+                });
+    }
+
+    /**
+     * This is when the "back" button is pressed
+     */
+    public void goBack(View view) {
+        Log.i(TAG + ":goBack", "Finish ingredient location edit");
+        Intent intent = new Intent(this, AddEditStoreIngredientActivity.class);
+        // make sure you can't go back to this activity
+        // https://stackoverflow.com/a/18957237
+        intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
+
+        setResult(Activity.RESULT_OK, intent);
+        finish();
+    }
+
+    /**
+     * Simplified method to make a simple snackbar
+     */
+    private void makeSnackbar(String msg) {
+        RecyclerView root = findViewById(R.id.location_listview);
+        Snackbar.make(root, msg, Snackbar.LENGTH_LONG)
+                .setAction("Ok", view1 -> {})
+                .show();
     }
 }
